+++ core/commons/networks/network.go
@@ -1,7 +1,0 @@
-package networks
-
-/*
-A marker interface for representing a generic test network. Developers should add their own interface that
-	extends this interface.
- */
-type Network interface{}