+++ core/commons/services/service.go
@@ -1,7 +1,0 @@
-package services
-
-/*
-Marker interface representing a generic service in a test network. The developer should implement their own
-	use-case-specific interface that inherits from this one.
- */
-type Service interface {}