+++ core/initializer/parallelism/parallel_test_params.go
@@ -2,7 +2,6 @@ package parallelism
 
 import (
 	"github.com/docker/distribution/uuid"
-	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 )
 
 /*
@@ -12,9 +11,6 @@ type ParallelTestParams struct {
 	// Name of the test to run
 	TestName            string
 
-	// Logic of the test to run
-	Test 				testsuite.Test
-
 	// Subnet mask that should be used for the Docker network that the test controller & network will run in
 	SubnetMask          string
 
@@ -22,6 +18,10 @@ type ParallelTestParams struct {
 	ExecutionInstanceId uuid.UUID
 }
 
-func NewParallelTestParams(testName string, test testsuite.Test, subnetMask string, executionInstanceId uuid.UUID) *ParallelTestParams {
-	return &ParallelTestParams{TestName: testName, Test: test, SubnetMask: subnetMask, ExecutionInstanceId: executionInstanceId}
+func NewParallelTestParams(testName string, subnetMask string, executionInstanceId uuid.UUID) *ParallelTestParams {
+	return &ParallelTestParams{
+		TestName: testName,
+		SubnetMask: subnetMask,
+		ExecutionInstanceId: executionInstanceId,
+	}
 }