+++ core/initializer/test_suite_runner/test_suite_runner.go
@@ -1,12 +1,13 @@
-package initializer
+package test_suite_runner
 
 import (
 	"encoding/binary"
 	"fmt"
 	"github.com/docker/distribution/uuid"
 	"github.com/docker/docker/client"
-	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
+	"github.com/kurtosis-tech/kurtosis/commons"
 	"github.com/kurtosis-tech/kurtosis/initializer/parallelism"
+	"github.com/kurtosis-tech/kurtosis/initializer/test_suite_metadata_acquirer"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"math"
@@ -20,54 +21,57 @@ const (
 	SUBNET_START_ADDR = "172.23.0.0"
 
 	BITS_IN_IP4_ADDR = 32
+
+	// TODO Pull this from the testsuite image!
+	networkWidthBits = 8
 )
 
 /*
 An executor to run one or more tests from a given test suite
  */
 type TestSuiteRunner struct {
-	// The test suite to run tests from
-	testSuite               testsuite.TestSuite
+	// Docker client to use for interacting with the Docker engine
+	dockerClient *client.Client
 
-	// The name of the Docker image that should be used to run the test controller container
-	testControllerImageName string
+	// The name of image of test suite
+	testSuiteImage               string
 
-	// Key-value mapping that will be passed as-is to the test controller container on startup in the form of Docker
+	// The name of the Kurtosis API Docker image
+	kurtosisApiImage string
+
+	// Key-value mapping that will be passed as-is to the test suite container on startup in the form of Docker
 	// 	environment variables
-	customTestControllerEnvVars map[string]string
+	customTestSuiteEnvVars map[string]string
 
 	// A string, meaningful only to the test controller, that represents the log level that the controller container should
 	//	run with
-	testControllerLogLevel	string
-
-	// The number of bits in a test network's subnet mask, such that 2 ^ this_value will be the maximum number of allowed
-	//  services in any given test network
-	networkWidthBits uint32
+	testSuiteLogLevel string
 }
 
 /*
 Creates a new TestSuiteRunner with the given parameters.
 
 Args:
+	client: Docker client to use when interacting with the Docker engine
 	testSuite: The test suite containing all the user's registered tests
 	testControllerImageName: The name of the Docker image of the test controller that will orchestrate test execution
-	testControllerLogLevel: The string representing the loglevel of the controller (the test suite runner won't be able
-		to parse this, so this should be meaningful to the controller image)
+	testSuiteLogLevel: The string representing the loglevel of the test suite (the initializer won't be able
+		to parse this, so this should be meaningful to the test suite image)
 	networkWidthBits: Each test will get a Docker network with a number of available IP addresses = 2^network_width_bits.
 		This parameter should be set high enough so that each test can fit all the services they want.
  */
 func NewTestSuiteRunner(
-			testSuite testsuite.TestSuite,
-			testControllerImageName string,
-			testControllerLogLevel string,
-			testControllerEnvVars map[string]string,
-			networkWidthBits uint32) *TestSuiteRunner {
+			dockerClient *client.Client,
+			testSuiteImage string,
+			kurtosisApiImage string,
+			testSuiteLogLevel string,
+			testControllerEnvVars map[string]string) *TestSuiteRunner {
 	return &TestSuiteRunner{
-		testSuite:                   testSuite,
-		testControllerImageName:     testControllerImageName,
-		testControllerLogLevel:      testControllerLogLevel,
-		customTestControllerEnvVars: testControllerEnvVars,
-		networkWidthBits:            networkWidthBits,
+		dockerClient:           dockerClient,
+		testSuiteImage:         testSuiteImage,
+		kurtosisApiImage:       kurtosisApiImage,
+		testSuiteLogLevel:      testSuiteLogLevel,
+		customTestSuiteEnvVars: testControllerEnvVars,
 	}
 }
 
@@ -84,28 +88,33 @@ Returns:
 		being retrieved. If this is non-nil, the allTestsPassed value is undefined!
  */
 func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParallelism uint) (allTestsPassed bool, executionErr error) {
-	allTests := runner.testSuite.GetTests()
+	stdoutDockerManager, err := commons.NewDockerManager(logrus.StandardLogger(), runner.dockerClient)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred creating the Docker manager")
+	}
 
-	// If the user doesn't specify any test names to run, run all of them
+	allTestNames, err := test_suite_metadata_acquirer.GetAllTestNamesInSuite(runner.testSuiteImage, stdoutDockerManager)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred getting the names of the tests in the test suite")
+	}
+
+	// If the user doesn't specify any test names to run, do all of them
 	if len(testNamesToRun) == 0 {
 		testNamesToRun = map[string]bool{}
-		for testName, _ := range allTests {
+		for testName := range allTestNames {
 			testNamesToRun[testName] = true
 		}
 	}
 
 	// Validate all the requested tests exist
-	testsToRun := make(map[string]testsuite.Test)
-	for testName, _ := range testNamesToRun {
-		test, found := allTests[testName]
-		if !found {
+	for testName := range testNamesToRun {
+		if _, found := allTestNames[testName]; !found {
 			return false, stacktrace.NewError("No test registered with name '%v'", testName)
 		}
-		testsToRun[testName] = test
 	}
 
 	executionInstanceId := uuid.Generate()
-	testParams, err := buildTestParams(executionInstanceId, testsToRun, runner.networkWidthBits)
+	testParams, err := buildTestParams(executionInstanceId, testNamesToRun, networkWidthBits)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred building the test params map")
 	}
@@ -119,12 +128,13 @@ func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParal
 	testExecutor := parallelism.NewTestExecutorParallelizer(
 		executionInstanceId,
 		dockerClient,
-		runner.testControllerImageName,
-		runner.testControllerLogLevel,
-		runner.customTestControllerEnvVars,
+		runner.kurtosisApiImage,
+		runner.testSuiteImage,
+		runner.testSuiteLogLevel,
+		runner.customTestSuiteEnvVars,
 		testParallelism)
 
-	logrus.Infof("Running %v tests with execution ID %v...", len(testsToRun), executionInstanceId.String())
+	logrus.Infof("Running %v tests with execution ID %v...", len(testNamesToRun), executionInstanceId.String())
 	allTestsPassed = testExecutor.RunInParallelAndPrintResults(testParams)
 	return allTestsPassed, nil
 }
@@ -135,7 +145,7 @@ Helper function to build, from the set of tests to run, the map of test params t
 Args:
 	testsToRun: A "set" of test names to run in parallel
  */
-func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]testsuite.Test, networkWidthBits uint32) (map[string]parallelism.ParallelTestParams, error) {
+func buildTestParams(executionInstanceId uuid.UUID, testNamesToRun map[string]bool, networkWidthBits uint32) (map[string]parallelism.ParallelTestParams, error) {
 	subnetMaskBits := BITS_IN_IP4_ADDR - networkWidthBits
 
 	subnetStartIp := net.ParseIP(SUBNET_START_ADDR)
@@ -155,14 +165,14 @@ func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]testsu
 
 	testIndex := 0
 	testParams := make(map[string]parallelism.ParallelTestParams)
-	for testName, test := range testsToRun {
+	for testName, _ := range testNamesToRun {
 		// Pick the next free available subnet IP, considering all the tests we've started previously
 		subnetIpInt := subnetStartIpInt + uint32(testIndex) * uint32(math.Pow(2, float64(networkWidthBits)))
 		subnetIp := make(net.IP, 4)
 		binary.BigEndian.PutUint32(subnetIp, subnetIpInt)
 		subnetCidrStr := fmt.Sprintf("%v/%v", subnetIp.String(), subnetMaskBits)
 
-		testParams[testName] = *parallelism.NewParallelTestParams(testName, test, subnetCidrStr, executionInstanceId)
+		testParams[testName] = *parallelism.NewParallelTestParams(testName, subnetCidrStr, executionInstanceId)
 		testIndex++
 	}
 	return testParams, nil