+++ core/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade new initializer to actually run a test suite!
 * Print rudimentary version of testsuite container logs
 * Refactor the new intializer's `main` method, which had become 550 lines long, into separate classes
+* Run tests in parallel
 
 # 0.9.0
 * Change ConfigurationID to be a string